Code Review
Install additional packages -  install.packages("knitr"), install.packages("devtools"), BiocManager::install("limma"), BiocManager::install("GEOquery"), BiocManager::install("ComplexHeatmap"),
The rest of the code runs. 
Based on this code, it is hard to understand what is the result and what the original question was. I also do not clearly understand the logic behind arranging the code in this order.